refactor

+++ src/cmd.h
@@ -77,7 +77,9 @@ inline bool matchFieldSep(StringRef ifs, char ch) {
  * @return
  * 0-255
  */
-inline int maskExitStatus(int64_t status) { return status & 0xFF; }
+inline int maskExitStatus(int64_t status) {
+  return static_cast<int>(static_cast<uint64_t>(status) & 0xFF);
+}
 
 } // namespace ydsh
